yolo5/requirements.txt
flask
pyyaml
loguru
detect
boto3
pymongo

# testing

pylint
pytest